requirements.txt
ruamel.yaml==0.15.94
python-dotenv==0.10.1
colorama=0.4.1
